requirements.txt
requests==2.32.4
streamlit==1.35.0
python-dateutil==2.9.0.post0